requirements-dev.txt
# Dev-only dependencies. NEVER bake into the production image.
# Install on the host:
#   pip install -r requirements-dev.txt
#   playwright install chromium
pytest>=8.0
pytest-playwright>=0.5.0
pytest-base-url>=2.1.0
Pillow>=10.0

# Runtime deps the app uses — needed on the host so unit tests that
# `from app import app` can collect without the container. These match
# what the Dockerfile installs for production.
flask>=2.0
requests>=2.28
